Refactor
Some refactor has been done

+++ SuCorrientazoADomicilioMarlonR/src/main/java/s4n/codechallenge/actorsdtos/commands/DroneCmd.java
@@ -11,7 +11,7 @@ import s4n.codechallenge.entities.Drone;
 @Setter
 @Builder
 public class DroneCmd {
-    private byte droneId;
+    private int droneId;
 
     public static Drone toModel(DroneCmd droneCmd) {
         return Drone.builder()
